refactor(oracle): extract scapy decode helpers module
